+++ libraries/MCP23008/test/unit_test_001.cpp
@@ -66,7 +66,10 @@ unittest(test_constructor)
   MCP23008 MCP(0x27);
 
   assertFalse(Wire.didBegin());
+
+  Wire.begin();
   MCP.begin();
+
   //  in fact invalid ...
   assertTrue(Wire.didBegin());
   assertTrue(MCP.isConnected());
@@ -78,6 +81,8 @@ unittest(test_lastError)
   Wire.resetMocks();
 
   MCP23008 MCP(0x27);
+
+  Wire.begin();
   MCP.begin();
   assertEqual(MCP23008_OK, MCP.lastError());